Added a link to letsencrypt-mosquitto
This docker container  makes it really easy to create encrypted connections to a Mosquitto MQTT server. It uses the caddy web server to generate lets encrypt certificates which the Mosquitto MQTT broker then uses.  It also includes the Mosquitto Go Authorization plugin https://github.com/iegomez/mosquitto-go-auth.   The idea is to make life as easy as possible for the Mosquitto administrator.